[*] Task2_Doctor
