--- Task1/Doctor.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Task2
-{
-    internal abstract class Doctor
-    {
-        public abstract void Heal();
-    }
-}
