--- Task1/Therapist.cs
@@ -1,26 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Task2
-{
-    internal class Therapist : Doctor
-    {
-        public override void Heal()
-        {
-            Console.WriteLine("Терапевт начал лечить...");
-        }
-    }
-}
-
-
-
-
-
-
-
-
-
-
